Can't access to Native\Electron\Concerns\LocatesPhpBinary
