+++ src/Traits/Installer.php
@@ -2,6 +2,8 @@
 
 namespace Native\Electron\Traits;
 
+use Symfony\Component\Process\Process;
+
 trait Installer
 {
     protected function installNPMDependencies($installer = 'npm')
@@ -22,17 +24,37 @@ trait Installer
     protected function installDependencies(?string $installer = null)
     {
         $installers = [
-            'npm'  => 'npm set progress=false && npm install',
+            'npm'  => 'npm install',
             'yarn' => 'yarn',
             'pnpm' => 'pnpm install',
         ];
 
-        if (!in_array($installer, array_keys($installers))) {
-            $this->error("Invalid installer {$installer} provided.");
-            $installer = $this->choice('Choose one of the following installers: npm, yarn, pnpm', array_keys($installers));
+        if (!array_key_exists($installer, $installers)) {
+            $this->error("Invalid installer ** {$installer} ** provided.");
+            $keys = array_keys($installers);
+            $techs = implode(', ', $keys);
+            $installer = $this->choice('Choose one of the following installers: ' . $techs, $keys, 0);
         }
 
-        $this->info("Installing {$installer} dependencies using the npm package manager...");
+        $this->info("Installing NPM dependencies using the {$installer} package manager...");
         $this->executeCommand("{$installers[$installer]}", $this->nativePhpPath());
+    }
+
+    protected function nativePhpPath()
+    {
+        return realpath(__DIR__ . '/../../resources/js');
+    }
+
+    protected function executeCommand($command, $path)
+    {
+        $process = (Process::fromShellCommandline($command, $path))->setTimeout(null);
+
+        if ('\\' !== DIRECTORY_SEPARATOR && file_exists('/dev/tty') && is_readable('/dev/tty')) {
+            $process->setTty(true);
+        }
+
+        $process->run(function ($type, $line) {
+            $this->output->write($line);
+        });
     }
 }
